requirements.txt
httpx==0.27.0
playwright==1.44.0
beautifulsoup4==4.12.3
praw==7.7.1
supabase==2.4.6
python-dotenv==1.0.1
tenacity==8.3.0
loguru==0.7.2
anyio==4.4.0
schedule==1.2.2
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
lxml==5.2.2
